refactor(modals): Replace CoreModal with UserFilesModal for file management in AssistantEditor and ProjectContextPanel
